% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calc.R
\name{age_calc}
\alias{age_calc}
\title{A Function to calculate one's age on a specified date}
\usage{
age_calc(dob, end_date = Sys.Date(), units = "years")
}
\arguments{
\item{dob}{requires a date formatted field.  For Example: as.Date(column,'\%m/\%d/\%Y')}

\item{end_date}{The date used to determine the age.  For Example: as.Date(column,'\%m/\%d/\%Y')}

\item{units}{expects: years,months,days}
}
\value{
A vector of numbers the same length as dob
}
\description{
A Function to calculate one's age on a specified date
}
