% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_path.R
\name{get_file_path}
\alias{get_file_path}
\title{A Function to create a fixed width file for the WTCS portal or IPEDS}
\usage{
get_file_path(file_pattern = "all", folder = getwd())
}
\arguments{
\item{file_pattern}{A partial filename as string.}

\item{folder}{folder location of the file to load}
}
\value{
A fully qualified path to one of more files
}
\description{
A Function to create a fixed width file for the WTCS portal or IPEDS
}
