% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getColType}
\alias{getColType}
\title{Internal functions used in write_fwf and read_matc_fwf that isolates file widths}
\usage{
getColType(lo, x)
}
\arguments{
\item{lo}{fwf_layout dataframe}

\item{x}{record id referenced in fwf_layout}
}
\description{
Internal functions used in write_fwf and read_matc_fwf that isolates file widths
}
\keyword{internal}
