% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fwf.R
\name{write_fwf}
\alias{write_fwf}
\title{A Function to create a fixed width file for the WTCS portal or IPEDS}
\usage{
write_fwf(x, output_file_name, type = "WTCS")
}
\arguments{
\item{x}{A dataframe with a column called record_id matching a layout file id}

\item{output_file_name}{The name of the export file}

\item{type}{use WTCS layout file or IPEDS layout file}
}
\value{
A boolean of TRUE
}
\description{
A Function to create a fixed width file for the WTCS portal or IPEDS
}
